requirements.txt
Flask
Flask-MySQLdb
Flask-WTF
passlib
Flask-Bootstrap
Flask-Compress
Flask-DebugToolbar
Flask-Login
Flask-Mail
Flask-Migrate
Flask-Moment
Flask-OpenID
Flask-PageDown
Flask-RESTful
Flask-Script
Flask-Session
Flask-SQLAlchemy
Flask-Uploads
